Don't remove item on key-press when input is focused

+++ src/app/pages/items/components/inventory/inventory.component.ts
@@ -56,9 +56,13 @@ export class InventoryComponent implements OnInit {
   }
 
   /**
-   * Removes the currently selected item from the inventor and clears the selection after
+   * Removes the currently selected item from the inventor and clears the selection IF there is no focused input-element
    */
   removeSelectedItem(): void {
+    // if there is any input-element that has focus we don't want to remove the item, since pressing Delete or Backspace would be intendet for editing the value of the input
+    const domInputeElements = Array.from(document.querySelectorAll<HTMLInputElement>('input'));
+    if (domInputeElements.some(element => element === document.activeElement)) return;
+
     if (this.indexToSelect !== -1) {
       this.characterService.removeItemFromInventory(this.indexToSelect);
       this.selectedItemService.setSelectedItem(-1);
